big refactor - visualization and processing seperated
